Fixed merge errors

+++ PHP/PHP/DeleteSale.cs
@@ -24,13 +24,28 @@ namespace PHP
             _pHPRepo = pHPRepo;
         }
 
-        private void button2_Click(object sender, EventArgs e)
+        private void RemoveButton_Click(object sender, EventArgs e)
         {
             int id;
-            int.TryParse(textBox1.Text, out id);
-            sale = _pHPRepo.GetSaleById(id);
-            _pHPRepo.RemoveSales(sale);
-            MessageBox.Show("Sucessful", "Sale has been deleted", MessageBoxButtons.OK, MessageBoxIcon.None);
+            if (int.TryParse(SaleID.Text, out id))
+            {
+                if (_pHPRepo.GetSaleById(id) != null)
+                {
+                    sale = _pHPRepo.GetSaleById(id);
+                    _pHPRepo.RemoveSales(sale);
+                    MessageBox.Show("Sucessful", "Sale has been deleted", MessageBoxButtons.OK, MessageBoxIcon.None);
+                    SaleID.Clear();
+                }
+                else
+                {
+                    MessageBox.Show("Could not find Sale: " + SaleID.Text);
+                }
+            }
+            else
+            {
+                MessageBox.Show("Please enter valid ID");
+            }
+
         }
 
         private void label1_Click(object sender, EventArgs e)
